Refactored views module
